feat: Edge AlertController 기능 분해 및 모듈화
